refactor: Move parsing lotto function to LottoParser
