+++ src/main/java/com/woowacourse/lotto/util/LottoParser.java
@@ -1,5 +1,6 @@
 package com.woowacourse.lotto.util;
 
+import com.woowacourse.lotto.domain.Lotto;
 import com.woowacourse.lotto.domain.LottoTicket;
 
 import java.util.Arrays;
@@ -7,6 +8,19 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class LottoParser {
+    public static Lotto parseStringToLotto(String input) {
+        String[] tokens = input.split(",");
+
+        List<Integer> lottoNumbers = Arrays.stream(tokens)
+                .map(String::trim)
+                .map(Integer::parseInt)
+                .collect(Collectors.toList());
+
+        IntendedLottoGenerator intendedLottoGenerator = new IntendedLottoGenerator(lottoNumbers);
+
+        return new Lotto(intendedLottoGenerator);
+    }
+
     public static LottoGenerator parseLottoGenerator(String lottoInput) {
         String[] tokens = lottoInput.split(",");
         List<Integer> list = Arrays.stream(tokens)
